+++ backend/canisters/local_user_index/api/src/lib.rs
@@ -35,7 +35,6 @@ pub enum Event {
     UserPrincipalUpdated(UpdateUserPrincipalArgs),
     DeleteUser(DeleteUser),
     SecretKeySet(Vec<u8>),
-    ChitEarned(ChitEarned),
     NotifyUniqueHumanProof(UserId, UniquePersonProof),
 }