+++ backend/canisters/local_user_index/impl/src/jobs/mod.rs
@@ -8,7 +8,7 @@ pub mod topup_canister_pool;
 pub mod upgrade_canisters;
 
 pub(crate) fn start(state: &RuntimeState) {
-    delete_users::start_job_if_required(state);
+    delete_users::start_job_if_required(state, None);
     make_btc_miami_payments::start_job_if_required(state);
     sync_events_to_user_canisters::start_job_if_required(state);
     sync_events_to_user_index_canister::start_job_if_required(state);