+++ backend/canisters/user_index/impl/src/lifecycle/post_upgrade.rs
@@ -6,6 +6,7 @@ use canister_tracing_macros::trace;
 use ic_cdk::post_upgrade;
 use stable_memory::get_reader;
 use tracing::info;
+use types::CanisterId;
 use user_index_canister::post_upgrade::Args;
 use utils::cycles::init_cycles_dispenser_client;
 
@@ -33,5 +34,30 @@ fn post_upgrade(args: Args) {
                 .push_back((user.principal, Some(user.user_id)));
         }
         crate::jobs::sync_users_to_identity_canister::start_job_if_required(state);
+
+        for canister_id in state.data.deleted_users.iter().map(|u| CanisterId::from(u.user_id)) {
+            let local_user_index = local_user_index_canister(canister_id, state.data.test_mode);
+            let event = local_user_index_canister::Event::AddCanisterToPool(canister_id);
+            state.data.user_index_event_sync_queue.push(local_user_index, event);
+        }
+        crate::jobs::sync_events_to_local_user_index_canisters::start_job_if_required(state);
     })
+}
+
+fn local_user_index_canister(canister_id: CanisterId, test_mode: bool) -> CanisterId {
+    let bytes = canister_id.as_slice();
+    if bytes > [0, 0, 0, 0, 2, 32, 0, 0, 1, 1].as_slice() && bytes < [0, 0, 0, 0, 2, 48, 0, 0, 1, 1].as_slice() {
+        return if test_mode {
+            CanisterId::from_text("pecvb-tqaaa-aaaaf-bhdiq-cai").unwrap()
+        } else {
+            CanisterId::from_text("nq4qv-wqaaa-aaaaf-bhdgq-cai").unwrap()
+        };
+    }
+    if bytes > [0, 0, 0, 0, 0, 160, 0, 0, 1, 1].as_slice() && bytes < [0, 0, 0, 0, 0, 176, 0, 0, 1, 1].as_slice() {
+        return CanisterId::from_text("aboy3-giaaa-aaaar-aaaaq-cai").unwrap();
+    }
+
+    assert!(test_mode);
+    // This will only be reached during tests + local development
+    CanisterId::from_text("be2us-64aaa-aaaaa-qaabq-cai").unwrap()
 }